+++ tests/function/test_record.py
@@ -17,16 +17,18 @@
 # specific language governing permissions and limitations
 # under the License.
 
+import decimal
 import os
 import sys
 import time
-import decimal
 
 from six.moves import configparser
 
 from datahub import DataHub
-from datahub.exceptions import ResourceExistException, InvalidOperationException
+from datahub.exceptions import ResourceExistException, InvalidOperationException, InvalidParameterException, \
+    ResourceNotFoundException
 from datahub.models import RecordSchema, FieldType, TupleRecord, BlobRecord, CursorType, CompressFormat
+from datahub.utils import to_binary
 
 current_path = os.path.split(os.path.realpath(__file__))[0]
 root_path = os.path.join(current_path, '../..')
@@ -90,7 +92,8 @@ class TestRecord:
 
         record_schema = RecordSchema.from_lists(
             ['bigint_field', 'string_field', 'double_field', 'bool_field', 'time_field', 'decimal_field'],
-            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP, FieldType.DECIMAL],
+            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP,
+             FieldType.DECIMAL],
             [False, True, True, True, True, True])
 
         print(TupleRecord(schema=record_schema))
@@ -151,6 +154,7 @@ class TestRecord:
             print(failed_records)
             assert len(failed_records) == 0
 
+            time.sleep(2)
             # ======================= get record =======================
             cursor = dh.get_cursor(project_name, topic_name, '0', CursorType.OLDEST)
             record_result = dh.get_tuple_records(project_name, topic_name, '0', record_schema, cursor.cursor, 4)
@@ -170,7 +174,8 @@ class TestRecord:
 
         record_schema = RecordSchema.from_lists(
             ['bigint_field', 'string_field', 'double_field', 'bool_field', 'time_field', 'decimal_field'],
-            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP, FieldType.DECIMAL],
+            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP,
+             FieldType.DECIMAL],
             [False, True, True, True, True, True])
 
         try:
@@ -248,7 +253,8 @@ class TestRecord:
 
         record_schema = RecordSchema.from_lists(
             ['bigint_field', 'string_field', 'double_field', 'bool_field', 'time_field', 'decimal_field'],
-            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP, FieldType.DECIMAL],
+            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP,
+             FieldType.DECIMAL],
             [False, True, True, True, True, True])
 
         try:
@@ -326,7 +332,8 @@ class TestRecord:
 
         record_schema = RecordSchema.from_lists(
             ['bigint_field', 'string_field', 'double_field', 'bool_field', 'time_field', 'decimal_field'],
-            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP, FieldType.DECIMAL],
+            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP,
+             FieldType.DECIMAL],
             [False, True, True, True, True, True])
 
         try:
@@ -403,8 +410,10 @@ class TestRecord:
         topic_name = "record_test_t%d_1" % int(time.time())
 
         record_schema = RecordSchema.from_lists(
-            ['bigint_field', 'string_field', 'double_field', 'bool_field', 'time_field'],
-            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP])
+            ['bigint_field', 'string_field', 'double_field', 'bool_field', 'time_field', 'decimal_field'],
+            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP,
+             FieldType.DECIMAL],
+            [False, True, True, True, True, True])
 
         try:
             dh_pb.create_project(project_name, '')
@@ -422,7 +431,9 @@ class TestRecord:
             # ======================= put record =======================
             # put tuple records
             failed_records = []
-            record = TupleRecord(schema=record_schema, values=[99, 'yc1', 10.01, True, 1455869335000000])
+            record = TupleRecord(schema=record_schema,
+                                 values=[99, 'yc1', 10.01, True, None,
+                                         decimal.Decimal('12.2219999999999995310417943983338773250579833984375')])
 
             # write by partition key
             record.partition_key = 'TestPartitionKey'
@@ -442,7 +453,8 @@ class TestRecord:
 
             # test failed records
             record1 = TupleRecord(schema=record_schema)
-            record1.values = [99, 'yc1', 10.01, True, 1455869335000000]
+            record1.values = [99, 'yc1', 10.01, True, 1455869335000000,
+                              decimal.Decimal('12.2219999999999995310417943983338773250579833984375')]
             record1.shard_id = '-1'
             record1.put_attribute('a', 'b')
 
@@ -461,7 +473,7 @@ class TestRecord:
 
             # ======================= get record =======================
             cursor = dh_pb.get_cursor(project_name, topic_name, '0', CursorType.OLDEST)
-            record_result = dh_pb.get_tuple_records(project_name, topic_name, '0', record_schema, cursor.cursor, 3)
+            record_result = dh_pb.get_tuple_records(project_name, topic_name, '0', record_schema, cursor.cursor, 6)
             print(record_result)
             assert record_result.record_count == 2
             assert record_result.records[0].values == record.values
@@ -670,7 +682,8 @@ class TestRecord:
 
             # ======================= get record =======================
             cursor = dh_pb_deflate.get_cursor(project_name, topic_name, '0', CursorType.OLDEST)
-            record_result = dh_pb_deflate.get_tuple_records(project_name, topic_name, '0', record_schema, cursor.cursor, 3)
+            record_result = dh_pb_deflate.get_tuple_records(project_name, topic_name, '0', record_schema, cursor.cursor,
+                                                            3)
             print(record_result)
             assert record_result.record_count == 3
             assert record_result.records[0].values == record.values
@@ -1025,6 +1038,204 @@ class TestRecord:
             clean_topic(dh_pb_deflate, project_name)
             dh_pb_deflate.delete_project(project_name)
 
+    def test_put_tuple_records_by_shard_id_pb(self):
+        project_name = "record_test_p%d_3" % int(time.time())
+        topic_name = "record_test_t%d_3" % int(time.time())
+
+        record_schema = RecordSchema.from_lists(
+            ['bigint_field', 'string_field', 'double_field', 'bool_field', 'time_field', 'decimal_field'],
+            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP,
+             FieldType.DECIMAL],
+            [False, True, True, True, True, True])
+
+        try:
+            dh_pb.create_project(project_name, '')
+        except ResourceExistException:
+            pass
+
+        # make sure project wil be deleted
+        try:
+            try:
+                dh_pb.create_tuple_topic(project_name, topic_name, 3, 7, record_schema, '1')
+                dh_pb.wait_shards_ready(project_name, topic_name)
+            except ResourceExistException:
+                pass
+
+            # ======================= put record =======================
+            record = TupleRecord(schema=record_schema,
+                                 values=[99, 'yc1', 10.01, True, 1455869335000000,
+                                         decimal.Decimal('12.2219999999999995310417943983338773250579833984375')])
+
+            dh_pb.put_records_by_shard(project_name, topic_name, "0", [record, record, record])
+
+            record1 = TupleRecord(schema=record_schema)
+            record1.set_value('bigint_field', 2)
+            record1.set_value('string_field', 'yc2')
+            record1.set_value('double_field', None)
+            record1.set_value(3, False)
+            record1.set_value(4, 1455869335000011)
+            record1.set_value(5, decimal.Decimal('12.2219999999999995310417943983338773250579833984375'))
+            record1.attributes = {'key': 'value'}
+
+            dh_pb.put_records_by_shard(project_name, topic_name, "0", [record1, record1, record1])
+
+            # ======================= get record =======================
+            time.sleep(1)
+            cursor = dh_pb.get_cursor(project_name, topic_name, '0', CursorType.OLDEST)
+            record_result = dh_pb.get_tuple_records(project_name, topic_name, '0', record_schema, cursor.cursor, 6)
+            print(record_result)
+            assert record_result.record_count == 6
+
+        finally:
+            clean_topic(dh_pb, project_name)
+            dh_pb.delete_project(project_name)
+
+    def test_put_tuple_records_by_shard_id_pb_failed(self):
+        project_name = "record_test_p%d_4" % int(time.time())
+        topic_name = "record_test_t%d_4" % int(time.time())
+
+        record_schema = RecordSchema.from_lists(
+            ['bigint_field', 'string_field', 'double_field', 'bool_field', 'time_field', 'decimal_field'],
+            [FieldType.BIGINT, FieldType.STRING, FieldType.DOUBLE, FieldType.BOOLEAN, FieldType.TIMESTAMP,
+             FieldType.DECIMAL],
+            [False, True, True, True, True, True])
+
+        try:
+            dh_pb.create_project(project_name, '')
+        except ResourceExistException:
+            pass
+
+        # make sure project wil be deleted
+        try:
+            try:
+                dh_pb.create_tuple_topic(project_name, topic_name, 3, 7, record_schema, '1')
+                dh_pb.wait_shards_ready(project_name, topic_name)
+            except ResourceExistException:
+                pass
+
+            # ======================= put record =======================
+            record = TupleRecord(schema=record_schema,
+                                 values=[99, 'yc1', 10.01, True, 1455869335000000,
+                                         decimal.Decimal('12.2219999999999995310417943983338773250579833984375')])
+
+            wrong_record_schema = RecordSchema.from_lists(
+                ['bigint_field', 'string_field', 'double_field', 'bool_field', 'time_field', 'decimal_field'],
+                [FieldType.STRING, FieldType.STRING, FieldType.STRING, FieldType.STRING, FieldType.STRING,
+                 FieldType.STRING])
+
+            wrong_record_schema_2 = RecordSchema.from_lists(
+                ['bigint_field', 'string_field', 'double_field', 'bool_field', 'time_field'],
+                [FieldType.STRING, FieldType.STRING, FieldType.STRING, FieldType.STRING, FieldType.STRING])
+
+            wrong_record = TupleRecord(schema=wrong_record_schema_2,
+                                       values=[99, 'yc1', 10.01, True, 1455869335000000])
+
+            wrong_record_2 = TupleRecord(schema=wrong_record_schema,
+                                         values=[99, 'yc1', 'a', True, 1455869335000000,
+                                                     '12.2219999999999995310417943983338773250579833984375'])
+
+            wrong_record_3 = TupleRecord(schema=wrong_record_schema,
+                                         values=['99', 'yc1', '10.01', 'true', '253402271999000001', '12.12'])
+
+            wrong_record_4 = TupleRecord(schema=wrong_record_schema,
+                                         values=['99', 'a', '10.01', 'true', '1455869335000000', '12.12'])
+
+            wrong_record_5 = TupleRecord(schema=wrong_record_schema,
+                                         values=['99', 'a', '10.01', 'true', '1455869335000000', '-'])
+
+            # ======================= invalid shard id =======================
+            try:
+                dh_pb.put_records_by_shard(project_name, topic_name, "-1", [record])
+            except ResourceNotFoundException as e:
+                assert e.error_msg == 'ShardId Not Exist. Invalid shard id:' + project_name + '/' + topic_name + '/-1'
+
+            # ======================= field size not match =======================
+            try:
+                dh_pb.put_records_by_shard(project_name, topic_name, "0", [wrong_record])
+            except InvalidParameterException as e:
+                assert e.error_msg == 'Record field size not match'
+
+            # ======================= type error =======================
+            try:
+                dh_pb.put_records_by_shard(project_name, topic_name, "0", [wrong_record_2])
+            except InvalidParameterException as e:
+                assert e.error_msg == 'Cannot cast empty string to d'
+
+            # ======================= project not existed =======================
+            try:
+                dh_pb.put_records_by_shard('a', topic_name, "0", [record])
+            except InvalidParameterException as e:
+                assert e.error_msg == 'Project name is missing or invalid:a'
+
+            # ======================= topic not existed =======================
+            try:
+                dh_pb.put_records_by_shard(project_name, 'a', "0", [record])
+            except ResourceNotFoundException as e:
+                assert e.error_msg == 'The specified topic name does not exist.'
+
+            # ======================= invalid timestamp =======================
+            try:
+                dh_pb.put_records_by_shard(project_name, topic_name, "0", [wrong_record_3])
+            except InvalidParameterException as e:
+                assert e.error_msg == 'Timestamp field value over range: 253402271999000001'
+
+            # ======================= invalid string length =======================
+            try:
+                wrong_record_4._values[1] = 'a' * (1024 * 1024 + 1)
+                dh_pb.put_records_by_shard(project_name, topic_name, "0", [wrong_record_4])
+            except InvalidParameterException as e:
+                assert e.error_msg == 'String field length: 1048577 exceed max length: 1048576'
+
+            # ======================= invalid string length =======================
+            try:
+                dh_pb.put_records_by_shard(project_name, topic_name, "0", [wrong_record_5])
+            except InvalidParameterException as e:
+                assert e.error_msg == 'Decimal field invalid: -'
+        finally:
+            clean_topic(dh_pb, project_name)
+            dh_pb.delete_project(project_name)
+
+    def test_put_blob_records_by_shard_id_pb(self):
+        project_name = "record_test_p%d_4" % int(time.time())
+        topic_name = "record_test_t%d_4" % int(time.time())
+
+        try:
+            dh_pb.create_project(project_name, '')
+        except ResourceExistException:
+            pass
+
+        # make sure project wil be deleted
+        try:
+            try:
+                dh_pb.create_blob_topic(project_name, topic_name, 3, 7, '')
+                dh_pb.wait_shards_ready(project_name, topic_name)
+            except ResourceExistException:
+                pass
+
+            # ======================= put record =======================
+            # put blob record
+            data = to_binary('blob data')
+
+            records = []
+
+            record0 = BlobRecord(blob_data=data)
+            record0.shard_id = '0'
+            record0.put_attribute('a', 'b')
+            records.append(record0)
+
+            dh_pb.put_records_by_shard(project_name, topic_name, "0", records)
+
+            # ======================= get record =======================
+            time.sleep(1)
+            cursor = dh_pb.get_cursor(project_name, topic_name, '0', CursorType.OLDEST)
+            record_result = dh_pb.get_blob_records(project_name, topic_name, '0', cursor.cursor, 3)
+            print(record_result)
+            assert record_result.record_count == 1
+            assert record_result.records[0].blob_data == data
+        finally:
+            clean_topic(dh_pb, project_name)
+            dh_pb.delete_project(project_name)
+
 
 # run directly
 if __name__ == '__main__':
@@ -1048,3 +1259,8 @@ if __name__ == '__main__':
     test.test_put_get_blob_records_pb_lz4()
     test.test_put_get_blob_records_pb_zlib()
     test.test_put_get_blob_records_pb_deflate()
+
+    test.test_put_tuple_records_by_shard_id_pb()
+    test.test_put_tuple_records_by_shard_id_pb_failed()
+
+    test.test_put_blob_records_by_shard_id_pb()